+++ CHANGELOG.md
@@ -130,6 +130,7 @@
 - list name not normalised when imported (#253)
 - GUI ignores date span in list.html URI if query is blank (#346)
 - search phrase dropped from list.html URI if date span is yyyy-mm (#347)
+- stats.lua first/last dates don't always agree with visible mails (#350)
 
 ## CHANGES in 0.9b: